feat(aws s3): now it's possible to upload images in aws

+++ src/modules/products/services/DeleteProductService.ts
@@ -1,14 +1,28 @@
 import { inject, injectable } from 'tsyringe';
 import { IProductsRepository } from '@modules/products/repositories/IProductsRepository';
+import { IStorageProvider } from '@shared/container/providers/StorageProvider/models/IStorageProviders';
+import { AppError } from '@shared/errors/AppError';
 
 @injectable()
 export class DeleteProductService {
     constructor(
         @inject('ProductsRepository')
         private productsRepository: IProductsRepository,
+        @inject('StorageProvider')
+        private storageProvider: IStorageProvider,
     ) {}
 
     async execute(id: string): Promise<void> {
-        return await this.productsRepository.delete(id)
+        const product = await this.productsRepository.findById(id);
+
+        if (!product) {
+            throw new AppError('product not found');
+        }
+
+        product.photos.forEach(async photo => {
+            await this.storageProvider.deleteFile(photo.name);
+        });
+
+        await this.productsRepository.delete(id);
     }
 }
